The `Cache` is now scoped to `legacy_state_sub`
